Removed the dependency of Body_Torso from Body_Uncalled hit modifier
Re-added BodyHit_Torso to ddraw.ini.
Reverted set_bodypart_hit_modifier function to pre-3.7.4 behavior.

+++ sfall/FalloutEngine.h
@@ -321,7 +321,7 @@ extern DWORD* ptr_gmouse_current_cursor;
 extern DWORD* ptr_gmovie_played_list;
 extern DWORD* ptr_GreenColor;
 extern DWORD* ptr_gsound_initialized;
-extern DWORD* ptr_hit_location_penalty;
+extern long*  ptr_hit_location_penalty;
 extern DWORD* ptr_holo_flag;
 extern DWORD* ptr_holopages;
 extern DWORD* ptr_hot_line_count;
